+++ tools/CicFEAlignment.h
@@ -41,7 +41,7 @@ class CicFEAlignment : public Tool
     void AlignInputs();
     void SetStaticPhaseAlignment();
     bool PhaseAlignment(uint16_t pWait_us = 10, uint32_t pNTriggers = 500);
-    bool WordAlignment(uint16_t pWait_us = 10);
+    bool WordAlignment(uint32_t pWait_us = 10);
     bool Bx0Alignment(uint8_t pFe = 0, uint8_t pLine = 4, uint16_t pDelay = 1, uint16_t pWait_ms = 100, int cNrials = 3);
     bool SetBx0Delay(uint8_t pDelay = 8, uint8_t pStubPackageDelay = 3);
     bool BackEndAlignment();
